provide .zprofile for non-interactive zsh
zsh is the default shell on macOS Catalina, which will be run non-interactively (likely for system commands), without running ~/.zshrc while startup. Without ~/.zshrc, which sources my custom general ~/.shell_profile, the shell will be lacking some necessary setup, such as $PATH and other environment variables.

The solution here is to provide a .zprofile, which will be run for all login shells, that detects if the shell is non-interactive to at least load ~/.shell_profile for setting up the basic shell environment.

Ref:
 - https://support.apple.com/kb/HT208050.
 - http://zsh.sourceforge.net/Guide/zshguide02.html